requirements.txt
annotated-types==0.7.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.10.0
attrs==25.4.0
cffi==2.0.0
Cython==3.1.6
Django==4.1.13
django-cors-headers==3.14.0
djangorestframework==3.14.0
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mysqlclient==2.2.7
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.28.0
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
